grouping/requirements.txt
pycrypto
redis
flask
flask-bootstrap
whoosh
